updated and fixed changes.
Works now! :)
